Update the staging and production APIClient tests to use the specified server, rather than always using the development server
Previously the baseURL would only be updated in setUp, which is called before each test. So, the server was always the default for all tests, which is development, even when it was intended to be production or staging. Now, in the staging and production test suites, we update both the server name and the baseURL to use the new server name and the tests run as expected against the specified server.