Fix bug https://trello.com/c/k9s5xryj
Fix bug caused by regression in commit 12857a89f95bcc11dd63cb1823bcd32db8b53615. The default numberOfRows for sections is 0. The numberOfRows for sections for this cell for an account with one group used to be 1, but, the aforementioned commit removed that, except for the HealthKit section, thus resulting in now content showing for other sections, notably Logout and Version.