Add support to enable/disable logging in production (disabled by default)
Change the "Server" alert to "Settings". Add an action button to the alert to "Disable logging" if logging is enabled, or to "Enable logging until next launch" if logging is disabled. Logging, if enabled, is verbose.
NOTE: Recall that this Debug UI is enabled by tapping four corners on the login view